# api/openapi.yaml
openapi: 3.1.0
info:
  title: OneImg API
  version: 1.0.0
  description: Breaking v1 API for OneImg. Legacy /api routes are not supported.
servers:
  - url: /api/v1
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [Account]
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    patch:
      operationId: updateMe
      tags: [Account]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateMeRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/DataResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
  /users:
    get:
      operationId: listUsers
      tags: [Users]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/Order' }
        - { name: id, in: query, schema: { type: integer, minimum: 1 } }
        - { name: username, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: integer, enum: [1, 3] } }
      responses:
        '200': { $ref: '#/components/responses/UserCollectionResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    post:
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateUserRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: oneimg-session }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, schema: { type: string } }
    Order: { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
...
  responses:
    ProblemResponse:
      description: RFC 9457 Problem Details
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    DataResponse:
      description: Successful response
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/DataEnvelope' } } }
...
    UserResponse:
      description: User response
      headers: { X-Request-ID: { schema: { type: string } }, Location: { schema: { type: string } } }
      content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/User' } } } } }
    UserCollectionResponse:
      description: Paginated user response
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { type: object, required: [data, meta], properties: { data: { type: array, items: { $ref: '#/components/schemas/ManagedUser' } }, meta: { $ref: '#/components/schemas/Meta' } } } } }
...
  schemas:
    DataEnvelope:
      type: object
      required: [data]
      properties: { data: {}, meta: { $ref: '#/components/schemas/Meta' } }
    Meta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, page_size, total, total_pages]
          properties: { page: { type: integer }, page_size: { type: integer }, total: { type: integer, format: int64 }, total_pages: { type: integer, format: int64 } }
    FieldProblem:
      type: object
      required: [field, code, message]
      properties: { field: { type: string }, code: { type: string }, message: { type: string } }
    Problem:
      type: object
      required: [type, title, status, detail, code, instance, request_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors: { type: array, items: { $ref: '#/components/schemas/FieldProblem' } }
    Permission:
      type: object
      required: [codes, bucket_ids]
      properties: { codes: { type: array, items: { type: string } }, bucket_ids: { type: array, items: { type: integer, minimum: 1 } } }
    User:
      type: object
      required: [id, username, role, permission, created_at, updated_at]
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: integer, enum: [1, 3] }
        permission: { $ref: '#/components/schemas/Permission' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ManagedUser:
      allOf:
        - { $ref: '#/components/schemas/User' }
        - { type: object, required: [passkey_count], properties: { passkey_count: { type: integer, format: int64, minimum: 0 } } }
...
    UpdateMeRequest:
      type: object
      required: [current_password]
      properties: { current_password: { type: string, format: password }, username: { type: string }, password: { type: string, format: password } }
...
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties: { username: { type: string, minLength: 3, maxLength: 50 }, password: { type: string, minLength: 6, maxLength: 100 }, role: { type: integer, enum: [1, 3] } }
...
security:
  - cookieAuth: []
  - bearerAuth: []